+++ byteIO/src/main/java/epam/rome/byteio/Main.java
@@ -1,0 +1,16 @@
+package epam.rome.byteio;
+
+import java.util.Map;
+import java.util.HashMap;
+
+public class Main {
+
+    public static void main(String[] args) {
+        String keywordsFilePath = "byteIO/src/main/java/epam/rome/byteio/keywords.txt";
+        ByteIOFindKeyWords.pullKeyWords(keywordsFilePath);
+        String readingFilePath = "byteIO/src/main/java/epam/rome/byteio/reading.class";
+        ByteIOFindKeyWords.readFile(readingFilePath);
+        
+        return;
+    }
+}
