+++ byteIO/src/main/java/epam/rome/byteio/reading.class
